+++ openbox/configuration_builder/configuration_builder_exceptions.py
